+++ crates/compiler/tests/fixtures/css/modules/global/output.js
@@ -55,13 +55,13 @@ farmModuleSystem.bootstrap();
 var entry = farmModuleSystem.require("b5d64806");
 
 
-//a1b6e7b5.css:
-  .base {
+//656e17cc.css:
+ .base {
   font-size: 20px;
 }
- .hide {
+.hide {
   display: none;
 }
- .show {
+.show {
   display: block;
 }