+++ crates/core/src/config/mod.rs
@@ -1,6 +1,7 @@
 use std::collections::HashMap;
 
 use serde::{Deserialize, Serialize};
+use swc_css_prefixer::options::Targets;
 use swc_ecma_ast::EsVersion;
 use swc_ecma_parser::{EsConfig, TsConfig};
 
@@ -121,16 +122,40 @@ pub struct ScriptConfig {
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 #[serde(rename_all = "camelCase", default)]
-pub struct CssConfig {
-  pub modules: bool,
+pub struct CssModulesConfig {
+  /// The paths regex to match css modules
+  pub paths: Vec<String>,
   pub indent_name: String,
+}
+
+impl Default for CssModulesConfig {
+  fn default() -> Self {
+    Self {
+      paths: vec![String::from("\\.module\\.(css|less|sass|scss)$")],
+      indent_name: String::from("[name]-[hash]"),
+    }
+  }
+}
+
+#[derive(Debug, Serialize, Deserialize, Clone, Default)]
+#[serde(rename_all = "camelCase", default)]
+pub struct CssPrefixerConfig {
+  #[serde(skip_serializing)]
+  pub targets: Option<Targets>,
+}
+
+#[derive(Debug, Serialize, Deserialize, Clone)]
+#[serde(rename_all = "camelCase", default)]
+pub struct CssConfig {
+  pub modules: Option<CssModulesConfig>,
+  pub prefixer: Option<CssPrefixerConfig>,
 }
 
 impl Default for CssConfig {
   fn default() -> Self {
     Self {
-      modules: false,
-      indent_name: String::from("[name]-[hash]"),
+      modules: Some(Default::default()),
+      prefixer: Some(Default::default()),
     }
   }
 }