+++ crates/toolkit/src/lib.rs
@@ -24,6 +24,7 @@ pub use swc_css_minifier;
 pub use swc_css_modules;
 pub use swc_css_parser;
 pub use swc_css_visit;
+pub use swc_css_prefixer;
 
 pub use swc_html_codegen;
 pub use swc_html_minifier;