+++ examples/less/src/typings.d.ts
@@ -1,2 +1,4 @@
 declare module '*.svg';
 declare module '*.css';
+declare module '*.png';
+declare module '*.less';