+++ examples/vue/src/Main.vue
@@ -4,6 +4,9 @@
 
 <script lang="ts" setup>
 import Welcome from './components/Welcome.vue';
+fetch('/api')
+  .then((response) => response.json())
+  .then((json) => console.log(json));
 </script>
 
 <style>