+++ packages/core/src/server/index.ts
@@ -1,4 +1,5 @@
 import { readFileSync } from 'fs';
+import http from 'node:http';
 
 import Koa from 'koa';
 // import serve from 'koa-static';
@@ -13,13 +14,15 @@ import {
   NormalizedServerConfig,
   normalizeDevServerOptions
 } from '../config/index.js';
-import { hmr } from './middlewares/hmr.js';
 import { HmrEngine } from './hmr-engine.js';
-import { brandColor, Logger } from '../logger.js';
+import { brandColor, Logger } from '../utils/logger.js';
 import { join } from 'path';
 import { fileURLToPath } from 'url';
-import { lazyCompilation } from './middlewares/lazy-compilation.js';
-import { resources } from './middlewares/resources.js';
+import { hmrPlugin } from './middlewares/hmr.js';
+import { lazyCompilationPlugin } from './middlewares/lazy-compilation.js';
+import { resourcesPlugin } from './middlewares/resources.js';
+import { proxyPlugin } from './middlewares/proxy.js';
+import { openBrowser } from './openBrowser.js';
 
 /**
  * Farm Dev Server, responsible for:
@@ -28,36 +31,50 @@ import { resources } from './middlewares/resources.js';
  * * compile the project in dev mode and serve the production
  * * HMR middleware and websocket supported
  */
-export class DevServer {
+
+interface FarmServerContext {
+  config: UserServerConfig;
+  app: Koa;
+  server: http.Server;
+  compiler: Compiler;
+  logger: Logger;
+}
+
+interface ImplDevServer {
+  createFarmServer(options: UserServerConfig): void;
+  listen(): Promise<void>;
+  getCompiler(): Compiler;
+}
+
+export class DevServer implements ImplDevServer {
   private _app: Koa;
 
+  _context: FarmServerContext;
   ws: WebSocketServer;
   config: NormalizedServerConfig;
   hmrEngine?: HmrEngine;
+  server?: http.Server;
 
   constructor(
     private _compiler: Compiler,
     public logger: Logger,
     options?: UserServerConfig
   ) {
+    this.createFarmServer(options);
+  }
+
+  createFarmServer(options: UserServerConfig) {
     this.config = normalizeDevServerOptions(options, 'development');
     this._app = new Koa();
-
-    // this._app.use(serve(this._dist));
-    this._app.use(resources(this._compiler));
-
-    if (this.config.hmr) {
-      this.ws = new WebSocketServer({
-        port: this.config.hmr.port,
-        host: this.config.hmr.host
-      });
-      this._app.use(hmr(this));
-      this.hmrEngine = new HmrEngine(this._compiler, this, this.logger);
-    }
-
-    if (this._compiler.config.config.lazyCompilation) {
-      this._app.use(lazyCompilation(this));
-    }
+    this.server = http.createServer(this._app.callback());
+    this._context = {
+      config: this.config,
+      app: this._app,
+      server: this.server,
+      compiler: this._compiler,
+      logger: this.logger
+    };
+    this.resolvedFarmServerPlugins();
   }
 
   getCompiler(): Compiler {
@@ -65,6 +82,7 @@ export class DevServer {
   }
 
   async listen(): Promise<void> {
+    const { port, open, https, host } = this.config;
     const start = Date.now();
     // compile the project and start the dev server
     if (process.env.FARM_PROFILE) {
@@ -72,10 +90,30 @@ export class DevServer {
     } else {
       await this._compiler.compile();
     }
-
     const end = Date.now();
+    this.server.listen(port);
+    this.startDevLogger(start, end);
 
-    this._app.listen(this.config.port);
+    const protocol = https ? 'https' : 'http';
+    const hostname = host || 'localhost';
+
+    if (open) {
+      openBrowser(`${protocol}://${hostname}:${port}`);
+    }
+  }
+
+  private resolvedFarmServerPlugins() {
+    const resolvedPlugins = [
+      lazyCompilationPlugin,
+      hmrPlugin,
+      resourcesPlugin,
+      proxyPlugin
+    ];
+    // this._app.use(serve(this._dist));
+    resolvedPlugins.forEach((p) => p(this));
+  }
+
+  private startDevLogger(start: number, end: number) {
     const version = JSON.parse(
       readFileSync(
         join(fileURLToPath(import.meta.url), '../../../package.json'),
@@ -89,11 +127,11 @@ export class DevServer {
             width: 40
           })
         )}
-Version ${chalk.green.bold(version)}
-
-🔥 Ready on ${chalk.green.bold(
-          `http://localhost:${this.config.port}`
-        )} in ${chalk.green.bold(`${end - start}ms`)}.
+  Version ${chalk.green.bold(version)}
+  
+  🔥 Ready on ${chalk.green.bold(
+    `http://localhost:${this.config.port}`
+  )} in ${chalk.green.bold(`${end - start}ms`)}.
     `,
         {
           padding: 1,