+++ examples/script-entry/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @farmfe-examples/script-entry
 
+## 0.0.17
+
+### Patch Changes
+
+- Updated dependencies [d9fe509]
+  - @farmfe/core@0.7.3
+
+## 0.0.16
+
+### Patch Changes
+
+- Updated dependencies [659bc72]
+- Updated dependencies [e0521e5]
+- Updated dependencies [926c9cb]
+  - @farmfe/core@0.7.2
+
 ## 0.0.15
 
 ### Patch Changes